Refactor DemandController and DemandHander
